obu/requirements.txt
# SmartV2X-CP Ultra — OBU Dependencies
# ======================================
numpy>=1.24.0
torch>=2.0.0
pyyaml>=6.0
httpx>=0.24.0
pyserial>=3.5
onnxruntime>=1.15.0
